+++ backend/canisters/token_metrics/api/src/queries/init_target_ledger.rs
@@ -1,4 +1,0 @@
-use crate::types::ledger_indexer::InitLedgerArgs;
-
-pub type Args = InitLedgerArgs;
-pub type Response = String;