+++ backend/canisters/token_metrics/api/src/types/mod.rs
@@ -1,2 +1,3 @@
 pub mod ledger_indexer;
+pub mod timer_status;
 pub mod token_data;