+++ backend/canisters/token_metrics/impl/src/updates/mod.rs
@@ -1,3 +1,2 @@
-pub mod init_target_ledger;
 pub mod start_processing_timer;
 pub mod stop_all_timers;